+++ src/main/java/com/coddingSchool/model/QuestionType.java
@@ -1,0 +1,7 @@
+package com.coddingSchool.model;
+
+public enum QuestionType {
+    UNIQUE,
+    MULTIPLE,
+    TRUE_OR_FALSE
+}
