+++ src/main/java/com/coddingSchool/run/RunningHtml.java
@@ -1,0 +1,19 @@
+package com.coddingSchool.run;
+
+import com.coddingSchool.fileWriters.HtmlWriter;
+
+public class RunningHtml {
+
+    static String CATEGORY_CSV_PATH = "/home/guilherme/Documentos/Level Up/planilha-dados-escola - Categoria.csv";
+    static String SUBCATEGORY_CSV_PATH = "/home/guilherme/Documentos/Level Up/planilha-dados-escola - Subcategoria.csv";
+    static String COURSE_CSV_PATH = "/home/guilherme/Documentos/Level Up/planilha-dados-escola-curso(copia).csv";
+
+    public static void main(String[] args) throws Exception {
+
+        HtmlWriter.createFile(
+                CATEGORY_CSV_PATH,
+                SUBCATEGORY_CSV_PATH,
+                COURSE_CSV_PATH
+        );
+    }
+}
