+++ src/main/java/com/coddingSchool/validations/EstimatedTimeValidator.java
@@ -1,0 +1,10 @@
+package com.coddingSchool.validations;
+
+public class EstimatedTimeValidator {
+
+    public static void timeValidator(int time, String error) {
+        if (time < 1 || time > 20) {
+            throw new IllegalArgumentException(error);
+        }
+    }
+}
