+++ src/main/java/com/coddingSchool/validations/ObjectValidator.java
@@ -1,0 +1,10 @@
+package com.coddingSchool.validations;
+
+public class ObjectValidator {
+
+    public static void cantBeNull(Object object, String error) {
+            if (object==null) {
+                throw new NullPointerException(error);
+            }
+    }
+}
