--- src/main/java/school/QuestionType.java
@@ -1,7 +1,0 @@
-package school;
-
-public enum QuestionType {
-    UNIQUE,
-    MULTIPLE,
-    TRUE_OR_FALSE
-}
