--- src/main/java/validations/EstimatedTimeValidator.java
@@ -1,10 +1,0 @@
-package validations;
-
-public class EstimatedTimeValidator {
-
-    public static void timeValidator(int time, String error) {
-        if (time < 1 || time > 20) {
-            throw new IllegalArgumentException(error);
-        }
-    }
-}
