--- src/main/java/validations/ObjectValidator.java
@@ -1,10 +1,0 @@
-package validations;
-
-public class ObjectValidator {
-
-    public static void cantBeNull(Object object, String error) {
-            if (object==null) {
-                throw new NullPointerException(error);
-            }
-    }
-}
